all QTPs except QTP3 working, due to thinking that Black -> Blue is equivalent to Black -> Black or Blue -> Blue

+++ cap_touch.h
@@ -6,7 +6,7 @@
 
 /* NOT TESTED */
 
-uint8_t calibrate_cap_touch(void);
+uint32_t calibrate_cap_touch(void);
 
 #endif	/* CAP_TOUCH_H */
